Bundle GeSHi 1.0.7.17
This version is known to work with the pastebin, but doesn't seem
to be available as a Composer package.

The easiest path is to bundle the library in this last known version
state, pending an update to use a more recent GeSHI version, or a more
modern highlighter.

Library release date: 2007-01-24